Patterns/src/main/java/patterns/behavioral/command/factory/RouteFactory.java
package patterns.behavioral.command.factory;

import java.util.HashMap;
import java.util.Map;

public final class RouteFactory {
	private static Map<String, Router> routes = new HashMap<String, Router>();
	private static RouteFactory factory = new RouteFactory();

	private RouteFactory() {
		routes.put(PropertyRouter.KEY, new PropertyRouter());
		routes.put(AutoRouter.KEY, new AutoRouter());
	}

	public static RouteFactory getInstance() {
		return factory;
	}

	public Router getRouter(final String key) {
		return routes.get(key);
	}

}
